Language switcher as own component
